pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "LLMFactory"
version = "0.0.1"
description = "A unified factory pattern interface for multiple LLM inference providers with multimodal support"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "Apache-2.0"}
authors = [
    {name = "M Chimiste"}
]
keywords = ["llm", "ai", "machine-learning", "inference", "anthropic", "openai", "gemini", "ollama", "bedrock", "multimodal"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
]
dependencies = [
    "pydantic",
    "numpy",
    "ollama",
    "anthropic[bedrock]",
    "openai",
    "gemini-ai",
    "sentence-transformers",
    "transformers",
    "torch",
    "torchvision",
    "torchaudio",
    "llama-cpp-python",
    "Pillow",
    "python-dotenv",
    "boto3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/M-Chimiste/LLMFactory"
"Bug Tracker" = "https://github.com/M-Chimiste/LLMFactory/issues"
"Source Code" = "https://github.com/M-Chimiste/LLMFactory"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --strict-markers --tb=short --disable-warnings"

[tool.black]
line-length = 120
target-version = ['py311', 'py312']
include = '\.pyi?$'

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false